requirements.txt
ultralytics
streamlit
opencv-python
Pillow
numpy
roboflow